ref: create init module
